FoodieHeaven/requirement.txt
streamlit
mysql-connector-python
fuzzywuzzy
python-Levenshtein
